backend/fastapi/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pydantic-settings>=2.0.0
pydantic>=2.0.0
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-multipart
requests>=2.31.0
nltk>=3.8.1
email-validator>=2.1.0
httpx>=0.27.0
aiofiles>=23.1.0
types-aiofiles>=23.1.0
cachetools==5.3.3
types-cachetools>=5.3.0
types-requests>=2.31.0
slowapi>=0.1.9
python-json-logger>=2.0.7
redis>=5.0.0
redlock-py>=1.0.10
celery[redis]>=5.3.0
psutil>=5.9.0
redis[asyncio]>=5.0.0
fastapi-cache2>=0.2.0
authlib>=1.3.0

# Async database drivers
asyncpg>=0.29.0
aiosqlite>=0.19.0
greenlet>=3.0.0
pytest-asyncio>=0.21.0
pyzipper>=0.3.6
reportlab>=4.0.0
jinja2>=3.1.0
celery>=5.3.6
pgvector>=0.2.0
sentence-transformers>=2.2.2
torch>=2.0.0
openai>=1.0.0
numpy>=1.24.0
scipy>=1.10.0
psutil>=5.9.0
boto3>=1.28.0
cyclonedx-bom>=4.0.0

